Clean up some TODOs
